Handle case where no seeding jobs produce .log files

+++ Condor_run.R
@@ -437,9 +437,15 @@ get_return_values <- function(log_dir, log_file_names) {
   return_values <- c()
   return_value_regexp <- "\\(1\\) Normal termination \\(return value (\\d+)\\)"
   for (name in log_file_names) {
-    loglines <- readLines(path(log_dir, name))
-    return_value <- as.integer(str_match(tail(grep(return_value_regexp, loglines, value=TRUE), 1), return_value_regexp)[2])
-    return_values <- c(return_values, return_value)
+    tryCatch({
+        loglines <- readLines(path(log_dir, name))
+        return_value <- as.integer(str_match(tail(grep(return_value_regexp, loglines, value=TRUE), 1), return_value_regexp)[2])
+        return_values <- c(return_values, return_value)
+      },
+      error=function(cond) {
+        return_values <- c(return_values, NA)
+      }
+    )
   }
   return(return_values)
 }
@@ -821,15 +827,15 @@ cat("Waiting for bundle seeding to complete...\n")
 monitor(clusters)
 rm(clusters)
 
+# Check if any seeding jobs produced .log files
+if (!any(file_exists(path(log_dir, str_glue("_seed_{hostnames}.log"))))) {
+  message("None of the seeding jobs produced a .log file!")
+  file_delete(bundle_path)
+  stop(str_glue("Aborting, see https://github.com/iiasa/Condor_run_R/blob/master/troubleshooting.md#seeding-fails-or-jobs-go-on-hold-without-producing-matching-log-files for possible solutions."))
+}
+
 # Determine which seed jobs failed
-tryCatch(
-  return_values <- get_return_values(log_dir, lapply(hostnames, function(hostname) return(str_glue("_seed_{hostname}.log")))),
-  error=function(cond) {
-    file_delete(bundle_path)
-    message("Cannot read log files of seeding jobs!")
-    stop(str_glue("Aborting, see https://github.com/iiasa/Condor_run_R/blob/master/troubleshooting.md#seeding-fails-or-jobs-go-on-hold-without-producing-matching-log-files for possible solutions."))
-  }
-)
+return_values <- get_return_values(log_dir, lapply(hostnames, function(hostname) return(str_glue("_seed_{hostname}.log"))))
 err_file_sizes <-  lapply(hostnames, function(hostname) return(file_size(path(log_dir, str_glue("_seed_{hostname}.err")))))
 failed_seeds <- is.na(return_values) | return_values != 0 | err_file_sizes != 0
 rm(return_values, err_file_sizes)
@@ -843,7 +849,7 @@ if (any(failed_seeds)) {
   if (length(failed_seeds[failed_seeds == TRUE]) == 1) {
     warning(str_glue("A seeding job failed, will refrain from scheduling jobs on the affected execution point {hostnames[failed_seeds]}. Probably, this host is currently unavailable."))
   } else {
-    warning(str_glue("Seeding jobs failed, will refrain from scheduling jobs on the affected execution point {str_c(hostnames[failed_seeds], collapse=', ')}. Probably, these hosts are currently unavailable."))
+    warning(str_glue("Seeding jobs failed, will refrain from scheduling jobs on the affected execution points {str_c(hostnames[failed_seeds], collapse=', ')}. Probably, these hosts are currently unavailable."))
   }
   hostdoms <- hostdoms[!failed_seeds]
   hostnames <- hostnames[!failed_seeds]
@@ -864,7 +870,7 @@ for (hostname in hostnames) {
 if (length(hostnames) == 1) {
   cat(str_glue("Seeding done: execution point {hostnames} has received and cached the bundle.\n"))
 } else {
-  cat(str_glue("Seeding done: execution point {str_c(hostnames, collapse=', ')} have received and cached the bundle.\n"))
+  cat(str_glue("Seeding done: execution points {str_c(hostnames, collapse=', ')} have received and cached the bundle.\n"))
 }
 cat("\n")
 rm(hostnames)
